requirements.txt
asn1crypto==0.24.0
backports.ssl-match-hostname==3.5.0.1
cached-property==1.3.1
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.4.2
docker==2.5.1
docker-compose==1.17.1
docker-pycreds==0.2.1
dockerpty==0.4.1
docopt==0.6.2
enum34==1.1.6
Flask==1.0.2
Flask-MySQL==1.4.0
funcsigs==1.0.2
functools32==3.2.3.post2
gunicorn==19.9.0
idna==2.7
ipaddress==1.0.22
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
MarkupSafe==1.1.0
mock==2.0.0
mysql-connector==2.1.6
pbr==5.1.1
pycparser==2.19
pycrypto==2.6.1
pygobject==3.26.1
PyMySQL==0.9.2
pyOpenSSL==17.5.0
pyxdg==0.25
PyYAML==3.12
requests==2.18.4
SecretStorage==2.3.1
six==1.11.0
stevedore==1.30.0
texttable==0.9.1
urllib3==1.22
virtualenv==16.1.0
virtualenv-clone==0.4.0
virtualenvwrapper==4.8.2
websocket-client==0.44.0
Werkzeug==0.14.1
